controller:lib: Updated code to fix compiler warning messages. Removed unused notification_id from set_wait_for_next_cmd function. Removed unused msg_type variables. Fixed variable type mismatch.

+++ controller/lib/src/util_imp.h
@@ -73,7 +73,7 @@ namespace avdecc_lib
         /**
          * Convert command value to its corresponding ACMP command name.
          */
-        const char * STDCALL acmp_cmd_value_to_name(uint16_t cmd_value);
+        const char * STDCALL acmp_cmd_value_to_name(uint32_t cmd_value);
 
         /**
          * Convert ACMP command name to its corresponding command value.
